terraform/infra/requirements.txt
azure-storage-file-datalake==12.21.0